// src/module/sold-inventory/infrastructure/input/dtos/event/incoming-sold-vehicle-event.dto.ts
type Website = Readonly<{
  id: string;
  name: string;
  url: string;
  status: 'ACTIVE' | 'INACTIVE';
  createdAt: string;
  updatedAt: string;
}>;

type Vehicle = Readonly<{
  id: string;
  externalId: string;
  url: string;
  name: string;
  description: string | null;
  year: number;
  transmission: string | null;
  mileage: number;
  frontImage: string;
  images: string[] | null;
  location: string;
  condition: 'USED' | 'NEW';
  originalPrice: number;
  price: number;
  currency: 'USD' | 'PEN';
  createdAt: string;
  updatedAt: string;
  websiteId: string;
  status: 'ACTIVE' | 'INACTIVE';
  website: Website;
}>;

export type SoldVehicleEventDto = {
  soldVehicles: Vehicle[];
}